@@ -20,23 +20,23 @@
     "publish": "yarn && lerna run clean && lerna run build && lerna publish --registry=https://registry.npmjs.org/"
   },
   "devDependencies": {
-    "@types/jest": "^27.0.2",
+    "@types/jest": "^27.4.0",
     "@typescript-eslint/eslint-plugin": "^4.33.0",
     "@typescript-eslint/parser": "^4.33.0",
     "browserify": "^17.0.0",
     "create-hash": "^1.1.3",
     "documentation": "^13.2.5",
     "eslint": "7",
-    "eslint-config-prettier": "^8.3.0",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.27.0",
-    "jest": "^27.3.1",
+    "eslint-plugin-react": "^7.29.4",
+    "jest": "^27.5.1",
     "lerna": "3.22.1",
     "prettier": "^2.2.1",
-    "ts-jest": "^27.0.7",
-    "ts-node": "^10.3.0",
-    "typescript": "^4.4.4",
-    "uglify-js": "^3.14.3"
+    "ts-jest": "^27.1.3",
+    "ts-node": "^10.6.0",
+    "typescript": "^4.6.2",
+    "uglify-js": "^3.15.3"
   },
   "dependencies": {
     "js-levenshtein": "^1.1.6"
