@@ -1,4 +1,5 @@
 module.exports = [
+["polygon","(del)","DEL",18,"(DEL)","3044022054dda4cba0d696d3f0b4e76200ac4682f396721036a6feb806fa8d623317760e02207a5ebde40cdd28d947dc49371bdde1beb915d776ac0b6e4ed8077d4bb7d0af21","0xec6432B90e7fD4d9f872cc5C781f05B617DB861E",false,false],
 ["polygon","(pos)_compound","COMP",18,"(PoS) Compound","3045022100bc6c39d0831aabe4207836ccfef88f0331dfc8b764bc4980f6bfe0fd2546c5cf02201636c8b88bb165998f86ff398a3da68e9557b3ad6f6d7f33e99b162ba7a7b201","0x8505b9d2254A7Ae468c0E9dd10Ccea3A837aef5c",false,false],
 ["polygon","(pos)_dai_stablecoin","DAI",18,"(PoS) Dai Stablecoin","3045022100b74ee418eb7eef1cde0876e9589a6ba217a3318ad840bb797c68adb4ecc57bd4022014fc46f93500576aca05034e243d12f4a28e56b81a3606d4c99d617ebebadaf5","0x8f3Cf7ad23Cd3CaDbD9735AFf958023239c6A063",false,false],
 ["polygon","(pos)_tether_usd","USDT",6,"(PoS) Tether USD","3045022100a9588096900e77f2989386c862582b225e587ba063401e53e953b94d55a04698022033fa8853df43d6ca84b10561aff0d9d804a2d0a49684cdf084a86732e03209b7","0xc2132D05D31c914a87C6611C10748AEb04B58e8F",false,false],
@@ -14,24 +15,31 @@ module.exports = [
 ["polygon","aave_matic_market_wbtc","AMWBTC",8,"Aave Matic Market WBTC","3045022100a8d864178f5f332844e5b8945461efca72f4f4e8f2d82073cd53ae5e65460a490220205ddf05d10acbba4e663eb4dc191b67b06829340c903c4449b6935c2c6cf017","0x5c2ed810328349100A66B82b78a1791B101C9D61",false,false],
 ["polygon","aave_matic_market_weth","AMWETH",18,"Aave Matic Market WETH","3044022061b6b779525e3df37e6d8e40da163b3edca2f90cf5dbc1e4b6e52f0c956bbcbe022035e3069875ab5c79b7d65efa57e46973d00252cdc40a5deff3460f5e02e597b3","0x28424507fefb6f7f8E9D3860F56504E4e5f5f390",false,false],
 ["polygon","aave_matic_market_wmatic","AMWMATIC",18,"Aave Matic Market WMATIC","304402205a27811f2e1837d46fa0179ed90ec344bbf5c6af49c8dca044a9b4ae754742c502200e2bc57d5260bdc4171b4443c598b6e524915fb7c3021170086664763721c0fd","0x8dF3aad3a84da6b69A4DA8aeC3eA40d9091B2Ac4",false,false],
+["polygon","abachi","ABI",9,"Abachi","3045022100afe972f846b77c429242b2162920b6dcd2a55147707c47b55b3cd5f4f4a74fd80220087a2d578581b0b26f0c2256bc54cacb3d3c70f8820c765c10e78525f3574012","0x6d5f5317308C6fE7D6CE16930353a8Dfd92Ba4D7",false,false],
 ["polygon","adamant","ADDY",18,"Adamant","3044022058ba00c91c27c8a7073ffeb67cf705f3d7727d34452c5304ac0f0cab3ce428c302206ee94a0e832239d7c4fe771f7fe25acf688881889a64b132f55d6b1a2f2718f5","0xc3FdbadC7c795EF1D6Ba111e06fF8F16A20Ea539",false,false],
 ["polygon","aioz_network","AIOZ",18,"AIOZ Network","3045022100d7327187553ba84adc1bc9284238ac7c0068e5d4c17233a1e512eaa4033aee700220631bda671092d9a8996cd44ae39347ba0f3fb0c5d0df655d1dca6b914b4026b1","0xe2341718c6C0CbFa8e6686102DD8FbF4047a9e9B",false,false],
+["polygon","algebra","ALGB",18,"Algebra","3045022100ffc9b269048fe1268ea09a1eeda0a21f7a9b70c1013ad9e61d47f21950a73773022027caef903f25f2fc8fbe655b764566a988226a6a31db39b6fba999b3636ce7ed","0x0169eC1f8f639B32Eec6D923e24C2A2ff45B9DD6",false,false],
 ["polygon","aluna","ALN",18,"Aluna","3045022100ea8ba091c2a7cb77a4154ee6295cef650f31e3efffbb25617239f5a8d4c6d3d902207c1f5cb6d78051a01103f276e4d3d5016ed307eb604e27a6eb9c9fb27fed2e67","0xa8fcEe762642f156b5D757b6FabC36E06b6d4A1A",false,false],
 ["polygon","amp","AMP",18,"Amp","3045022100de0b2a62ac2b686b590e16f5ad304a55fbbb942cd8f1fdf143163a7bfa77cea0022029c0819b4bcd42965b7b44c3e4fa459ef52cdb6933facdae09e3c5244af068f3","0x0621d647cecbFb64b79E44302c1933cB4f27054d",false,false],
 ["polygon","ampleforth_governance","FORTH",18,"Ampleforth Governance","3045022100c5645d780a9d15b5ff91593ac851a778f0d639fa140b63a53a7ddb4146a2739202206ccce793ea9cc19c944fb55974788b13a494a95803ad736ee24fd3bb01f6ceb9","0x5eCbA59DAcc1ADc5bDEA35f38A732823fc3dE977",false,false],
 ["polygon","ankr_eth2_reward_bearing_certificate","AETHC",18,"Ankr Eth2 Reward Bearing Certificate","304402205626e4fc42d2da26e30d6df185804db141f95ba0bc9554a693b694d85c82539a02204a1833c7c05590c71326c02f90c57aa39db4d4221f0e1e5e9f4fcfecbd3d980b","0xC4e82BA0Fe6763cbE5E9CbCA0ba7cbD6F91C6018",false,false],
+["polygon","apreum","APU",18,"Apreum","3044022022f19f6a5a41ee8f6bcb397cb87cb2b1a552e9c89aaca48fdbbff8f1894c77a5022023832d7dfbafa7d3994554eacc43034c580aaaf74c88fc8fac95ed82f8156bfb","0xC155504787e9430180F33F35eDd7C5eC06Cd5761",true,true],
 ["polygon","armor","ARMOR",18,"Armor","3045022100af076c9ab6176715c9361059a33de9d83d7bb33a7c44487e638968fe2d3a099402206c7e70ba8d1bb5e8c0e52dfae1ca44d4c39b8af04c60fca94828d98396076805","0x5de4005155933c0e1612Ce808f12B4cd8DAabc82",false,false],
 ["polygon","augury_finance","OMEN",18,"Augury Finance","3045022100d21d6e7fd83bc5940537b858bc6c6acecae970e8b13501bf25e2b12acb022f1e02201965c2d4f24945dead8914dc2b34f2483f3b9040799f3d17b40dcad3339243aa","0x76e63a3E7Ba1e2E61D3DA86a87479f983dE89a7E",false,false],
+["polygon","aurusgold","AWG",18,"AurusGOLD","30450221009ad5aa83db25bc1477fdf25f38af01d421322e747931bb5233d3945b59f002d902203c0870405f19e1638ad40f58aaf8347e2773174bd20f2508a1363549ac01cf2b","0xAEe0ffb690B37449B7f1C49B199E1E3ec6084490",false,false],
 ["polygon","bac","BAC",18,"BAC","30440220530fab88d00262cce862e93dfa18f0390b70451d73b2ee8ad204b35403fbd985022009075a365756e0e12a8b189d02b7c136459afe05f7d8ecc6afc8654a0a2ccd33","0xcdE5E3D90631cacB842fA86f62dD0fc08A3ce43d",false,false],
 ["polygon","balancer","BAL",18,"Balancer","3044022065dd446f0899eaf96b76e325fcdbb44011aa2cc74c97c8efeaefcba268c8ef88022004b20ead8b2627dfdb27295fd1549390205ffdfb101259c5e52049012e1b5516","0x9a71012B13CA4d3D0Cdc72A177DF3ef03b0E76A3",false,false],
 ["polygon","bancor","BNT",18,"Bancor","304402204b1b00e8d2b713b42b1cac76c4977ba8b327da2e7edaeacdb208061dd28b214202204d3e78d072bfb6bb11a342f4764444bdee2cd5b1cb4e748f0425571dd7021fc9","0xc26D47d5c33aC71AC5CF9F776D63Ba292a4F7842",false,false],
 ["polygon","bandtoken","BAND",18,"BandToken","3044022075f96cea4bb8035072670571807593f8ef1ea06ab86aef4ad07ca3bee21ad1df022076c94c82bee2509c4fe0a7c9db819188ba3805567a9ecca3c98caac8740291e4","0xA8b1E0764f85f53dfe21760e8AfE5446D82606ac",false,false],
+["polygon","bankless_defi_innovation_index_(pos)","GMI",18,"Bankless DeFi Innovation Index (PoS)","304402205efce28d158faf0e725c1d9737bf576d8dadfb1a9bc94263cf52631bbd1858d202202167d371ca4e308d9c95aec7e1465a018b0d6eb1cc46f045c2393398121ed431","0x7FB27EE135dB455de5Ab1CcEC66a24cBC82E712D",false,false],
 ["polygon","bat","BAT",18,"BAT","3045022100cf93e17a4073c83ddab8b64e01c0c64d6c2cb96a87c3a84a537bc1905933762902205bf6f9e94ee53d75e7c3ee7381703eff8131befbedfd72ec6cbc60d2ff942009","0x3Cef98bb43d732E2F285eE605a8158cDE967D219",false,false],
 ["polygon","beefy.finance","BIFI",18,"beefy.finance","3045022100be4c0e7ec00fa9fd985b145dea6b86f89ea3284a17912b8de1c1238910f69ad402207b62b18577fc9d8acf3927f6fc3c109ca76eb66366e389e4a86d93d826211db7","0xFbdd194376de19a88118e84E279b977f165d01b8",false,false],
 ["polygon","betprotocoltoken","BEPRO",18,"BetProtocolToken","3045022100db4b16fb793cdfdd83067cbdff8cf6cc0552cb159ee8e0187174bef6fa2224f4022011ca25280fa0eff139fec6e0c76628258d319ff0b37abe24147c623751ea26a4","0x07CC1cC3628Cc1615120DF781eF9fc8EC2feAE09",false,false],
 ["polygon","binance-usd","BUSD",18,"binance-usd","30440220306d65bd1bf54b51405a55dcea267903367f453aae13e7216f6cebb23dcbb6e602203c279f533074b262ca907f0c1c7444ba9cd244154f3d655e03b7383dd346d07c","0xdAb529f40E671A1D4bF91361c21bf9f0C9712ab7",false,false],
 ["polygon","bitfinex_leo_token","LEO",18,"Bitfinex LEO Token","3044022054e6d7c9f16d4a6128edd523e016180098c29c0585bbe48ea410254e681952a502207c5a3860f2f295f2b07171b2e0724a3e806d654a378e08baedc422f5c7bd865d","0x06D02e9D62A13fC76BB229373FB3BBBD1101D2fC",false,false],
+["polygon","blackpool_token_(pos)","BPT",18,"BlackPool Token (PoS)","304402207f74e8cc4b32aafa132655f8da80366dce0d7ec60ba53eb595d9ada6388609d502200d12abff69a31fa70ea9afa896cc3b0065673b6be884ca2d87d226417f115076","0x6863BD30C9e313B264657B107352bA246F8Af8e0",false,false],
 ["polygon","blockv","VEE",18,"BLOCKv","3045022100d773ab7c886b3a27979a1bbbb8e9f74d4859c533304092ea8ce81de559be0c96022002fefc041758c7b9ffbf6870c876af5ba66f8ebc86af37552fa5562186c8c474","0xf1c1A3C2481A3a8A3f173A9ab5Ade275292A6fA3",false,false],
+["polygon","blok","BLOK",18,"BLOK","3045022100c163912ceda6b2f5ae6c7adf21f46dc85d282a914e08fba485fd24cef0af5274022072824adf7a17b1f200710c1f052bf3dce11a0eeed1c59faca05acf3c9e55574a","0x229b1b6C23ff8953D663C4cBB519717e323a0a84",false,false],
 ["polygon","bluzelle","BLZ",18,"Bluzelle","304402203b0e1be78acac09c1f4d661dd3dad5c291c2b0c107531be52687d1834ff9afd50220443f3ab75a8337c7d3cdaf857ef35cce24f0bbc82d83cea08713ca64c8346df8","0x438B28C5AA5F00a817b7Def7cE2Fb3d5d1970974",false,false],
 ["polygon","bnb","BNB",18,"BNB","3044022076d39107a36eaccd897219063fa00eed78903b739ff83bad07990d3f2122b81902203916c2533447e0b7bbfabb45e1dfce2440c664ec29500a6f4e571bf0f0b617a2","0x3BA4c387f786bFEE076A58914F5Bd38d668B42c3",false,false],
 ["polygon","bone_token_(boneswap)","BONE",18,"Bone Token (BoneSwap)","304402203a15aaeced1965f8a9d19a1ee716108e1c73f1f5705d766bdbd3490e7471d18402205cc59469cdf7e431eab6a0ab7f0ba4518f487d5a8216e7130b5ae496bac6c304","0x80244C2441779361F35803b8C711C6c8fC6054a3",false,false],
@@ -39,25 +47,41 @@ module.exports = [
 ["polygon","bounty0x","BNTY",18,"Bounty0x","304402206bb140450d162c3bf97a049dd4af1e09b9f245e88b79ec80896ddbbc321d707202200868a516eb569065c5158fc6214b56e70fb0aba000f42ff6e5c15f0e6bae7e6f","0x0f40D69Ff494a9e74b1F20A89EfF299B2A05918F",false,false],
 ["polygon","btc_2x_flexible_leverage_index","BTC2X-FLI",18,"BTC 2x Flexible Leverage Index","30440220260fe077037f2ea7283410a03b93ec0880fa4ea14e55ee5c36e0c017caa72fe4022059ecd804021753fb1fbe928c16f825b44654454c7d1d5d775dba8de0eea7491d","0x4b54Bc363f5F9c6E0fcd82EaC6919aE213464Cc6",false,false],
 ["polygon","btu_protocol","BTU",18,"BTU Protocol","304402200e7b49c9063a554ae4ce257e36df3da4afc6ce7ffc99d8c2c02419381475fe61022027aae3d6973f9cfd77ceebfd77e3fdcbb51f2d148a6801724d51b626f1cd7e3f","0xFdc26CDA2d2440d0E83CD1DeE8E8bE48405806DC",false,false],
+["polygon","bullieverse","BULL",18,"Bullieverse","30440220626cb9e03e72d10b6706c4424589c40f5bf8d37032e3a93cd2f621eae1de9b6302200e2176f986c49a50068e3c41dcc751efaabc6791540e4c3f159cc43b70a16738","0x9f95e17b2668AFE01F8fbD157068b0a4405Cc08D",false,false],
 ["polygon","bzx_protocol_token","BZRX",18,"bZx Protocol Token","304402206fa371d32fe05821a0bb28306c3f3e2f0cf2f1a6efec349bcef67894575459c8022025e11ef31b7be14dd464c05237267105df54691f27ed50a6d49c64f1d798f0b4","0x54cFe73f2c7d0c4b62Ab869B473F5512Dc0944D2",false,false],
+["polygon","cad_coin","CADC",18,"CAD Coin","3045022100cb5a2ea5265d4c542ad6cebda4dca9d9dd5712d1448cdbb85e578f0d5f835f9f02200716801fdd44b9066da50ecc70cf0a09a73557d61ba19948c0f24429ab55c196","0x5d146d8B1dACb1EBBA5cb005ae1059DA8a1FbF57",false,false],
 ["polygon","cardstack","CARD",18,"Cardstack","30440220692705b2b40824a4e54ea19f3c25d012bf770fbfceddb6dbad50b7d18c31193b02207df69bfea34b27d87a075503dcbc3fe5dc1c1274d80a95ab487dc5e406414aac","0x3A4BeD49Ab592A227bAe8F41eFd0d38E6e06d96c",false,false],
 ["polygon","cargox","CXO",18,"CargoX","30440220306051149039bf55bf4c3f68194216508b514becc225db631b1140470399a59c022046dd1317f9ccd33c076fcc3d9222373b548a3754bf9339b8f64a45e62fce541d","0xf2ae0038696774d65E67892c9D301C5f2CbbDa58",false,false],
 ["polygon","celsius","CEL",4,"Celsius","3044022073bf3843575fd64633ca39a68160cad26a12c1a6714e5c0ef3464ccf564c7e9802202f40c3359d904bd7623e41f374ec0e1aa34451239cf5468ff1f75a2eb25a22f4","0xD85d1e945766Fea5Eda9103F918Bd915FbCa63E6",false,false],
+["polygon","cerby_token","CERBY",18,"Cerby Token","304402200f2ea2ed631fff9298f1293cc976c4c27ad2af9f1de804d54ee700379d3e2ebb02201f6bbb1f7765f7aa60592e1c390449014a86eadc53ba9b03577bde85a6cc150b","0xdef1fac7Bf08f173D286BbBDcBeeADe695129840",false,false],
 ["polygon","chainlink_token","LINK",18,"ChainLink Token","304402207253a5dc3dd9edd9a3fbae8779c20bfff5e3853a6ee46d76903c4c27f7cc521502204f3355c7ff2ce78fd5ba38b13037cafd481a5c058635e649588c8e42eb5e36f7","0x53E0bca35eC356BD5ddDFebbD1Fc0fD03FaBad39",false,false],
 ["polygon","chiliz","CHZ",18,"chiliZ","3045022100e2b143261ef36b6118d1d1ed78dd4c7b93bf7b781a7dcbf4775533c1dd5b3af30220187790828d395732ab555173bfd676c9a6044bc8f48221a0338d5f3209f47e93","0xf1938Ce12400f9a761084E7A80d37e732a4dA056",false,false],
 ["polygon","clintex","CTI",18,"ClinTex","3044022004499c41cb7abd51c9aa9f92f1e9284c16f6cd717cc05db775b0425a535517e90220723894c31d8c429c2a4244b698d6cc9adc3b255329e9895aa217282b96e519e1","0x8Ba941b64901E306667a287A370F145d98811096",false,false],
+["polygon","conch_coin","CONCH",9,"CONCH COIN","30440220039a1d07936fea859ae07e77f123a15748d58f2f05cad2596df012feeab678b002206ee54bc8c5253393d35731019e3896ea78b4a1d4ba156bf6898904133b201904","0x7cC9E49E10c9924F74f90C09f8f30F1f055acdd1",false,false],
+["polygon","cresio","XCRE",18,"Cresio","30450221009d09ae90abde22b61ab29b2e4d2e5942af655b59af035c37b73a60c1e68da28202207b70156fbdc3f85f0d1a87c855cfab5f7e51888b21aeda77b78d68e0c9169869","0xFA3c05C2023918A4324fDE7163591Fe6BEBd1692",false,false],
+["polygon","crowdtoken","CROWD",18,"CrowdToken","304402201cbbc0133a3c3dee0a276bec3085df5d755b715f925de43a21b740fcaba72e42022078087ac47fff6ce0e33f677a05c135f59c4ed27954b263b1eb98db5ac91d925b","0x483dd3425278C1f79F377f1034d9d2CaE55648B6",false,false],
 ["polygon","crv_(pos)","CRV",18,"CRV (PoS)","3045022100f379c74676f646ea2c59b1ad896ebb81ddee85db0f1a7cd9163251a49f31367c0220466c4cbe449cc63e5a94b262aa7bcd6c8f6cfd99ef1fcc356f9f597ee4148b56","0x172370d5Cd63279eFa6d502DAB29171933a610AF",false,false],
 ["polygon","cryption_network_token","CNT",18,"Cryption Network Token","304402201e1bf3b5cff585b6d0ac0a92f55ee0c01130a08c5c07caa4969310c1fa300a1a02203347f607a4e38df75289ceffa07470356e079930fa2c54371c31954a8b0d4dd2","0xD1e6354fb05bF72A8909266203dAb80947dcEccF",false,false],
-["polygon","crypto.com_coin","CRO",8,"Crypto.com Coin","3044022016fb0d88052730549a29a2164ae51a5ae21b750f9906e0a5210510d66f4aba550220700aa5c692177bb2bd88f47c6c392e401edba868cc7d44de9a31c4b1d96fc5f2","0xAdA58DF0F643D959C2A47c9D4d4c1a4deFe3F11C",false,false],
+["polygon","crypto.com_coin","CRO",8,"Crypto.com Coin","3044022016fb0d88052730549a29a2164ae51a5ae21b750f9906e0a5210510d66f4aba550220700aa5c692177bb2bd88f47c6c392e401edba868cc7d44de9a31c4b1d96fc5f2","0xAdA58DF0F643D959C2A47c9D4d4c1a4deFe3F11C",false,true],
+["polygon","crypto_development_services","CDS",8,"Crypto Development Services","304402207418870351d4a4cca44b1441db2b321f034a52b2c2434aa90e3183995daa2489022050747189f082a03629f4e7450a92cdf62ed74f5155356732868d8552dd4303f5","0x23f07a1C03e7C6D0C88e0E05E79B6E3511073fD5",false,false],
+["polygon","data_economy_index_(pos)","DATA",18,"DATA Economy Index (PoS)","3045022100b9fd462a812a5686883f307b95982d7fb077456e989f4c813a6ac6cf959d6318022070b8112e42177dc22a54131691ac6b24a83909ba0c27b80c21714cf07b0a2a94","0x1D607Faa0A51518a7728580C238d912747e71F7a",false,false],
 ["polygon","decentr","DEC",18,"Decentr","3043022011d7b66a4a5ff05f0fb26a4757a203386d4391989b394635c6e9555d60050434021f09cf35aa21e77fd650c78054ca528f663804b005b8cec5806f72c5f92b1a2a","0x314338767151e6Ed23DB8A41c6F6943c1958Ee56",false,false],
+["polygon","decentral.games","$DG",18,"decentral.games","3044022055528abd7fd74f4689271456bfdeed4fdbddf93c40b726d95fbcb3952173e8a9022007a30afdaeabb781232958bb8846000703b6f470c907366a8cea510796225c67","0x2a93172c8DCCbfBC60a39d56183B7279a2F647b4",true,true],
+["polygon","decentral_games_(pos)","DG",18,"Decentral Games (PoS)","30440220668297dd6eebca1f1e0ca2b606991c41c144cf5dd34ba7cb27cbe2321b51993d022014c0bd1b3c70d5d7683ebce108c01626d755350a92c97bae86bc01f6cb05f4d3","0xef938b6da8576a896f6E0321ef80996F4890f9c4",false,false],
+["polygon","decentral_games_governance_(pos)","XDG",18,"Decentral Games Governance (PoS)","3045022100a727a73ef994ab66871157580b79a2d7e3ca060246ce732c33bd8b8ff53271e80220476dba9038f225520b9e9257ff73159b28ace0d39593a13e797b171695cdba3e","0xc6480Da81151B2277761024599E8Db2Ad4C388C8",false,false],
 ["polygon","decentraland","MANA",18,"Decentraland","30440220436699eef62d0e27cde881617e2bcc77e99fee8ddf20f7fd716bdfb49a2d5699022073b43e4d889526c69e07a87d377fc401986c2f620287c98396ee7e449e82062d","0xA1c57f48F0Deb89f569dFbE6E2B7f46D33606fD4",false,false],
+["polygon","defipulse_index_(pos)","DPI",18,"DefiPulse Index (PoS)","3045022100c3404c5ddbb4da1fcecb7820eda85ae85c8d980b54876597ad8f93c190f071080220622e8d816ff2d053cfc8bc59a329858d19b3b4290706dde80602fcea7f2d744f","0x85955046DF4668e1DD369D2DE9f3AEB98DD2A369",false,false],
+["polygon","derace_token","DERC",18,"DeRace Token","3045022100f7b8da95f5ef57071eaebe38cca8dec6e843547de0c595f80235b6c28b681a2d0220088b5f2ca301ef392049d312723dad859abbbfaf783a8dcf71372674642f479e","0xB35fcBCF1fD489fCe02Ee146599e893FDCdC60e6",false,false],
 ["polygon","deri","DERI",18,"Deri","3044022051e056f54d588094f90a33f8407931b7a93c6c2f36edd9ae1bfe8995339fe69f022061993acf220da2c692f13eaa075e8e77e6195d495704c45df7c0c075571f3cdb","0x3D1D2aFd191b165D140e3E8329E634665fFB0E5e",false,false],
+["polygon","deus","DEUS",18,"DEUS","3045022100d7908e152e7416bf5e31981e15bd108c77387bf592f2f8c56be4b681798a99bc022067729913b3d6e419519faa9a95f6e6ddc79969e8d44da712b30ef2867a1a3b8d","0xDE5ed76E7c05eC5e4572CfC88d1ACEA165109E44",false,false],
+["polygon","dfx_token_(pos)","DFX",18,"DFX Token (PoS)","3045022100f4a521690a92991e537c140c01e8e55a975b4e2ef6e66b4c31b4ba7897874df7022030dec285f9009065db7f294953aa1f4ae615331387024f2fc2092162e669f04f","0xE7804D91dfCDE7F776c90043E03eAa6Df87E6395",false,false],
 ["polygon","dfyn_token_(pos)","DFYN",18,"DFYN Token (PoS)","304402204903b3c6afa507aef4cd88fb080d92a0b55a5b230c987b1add3493736db56e2302202e4b03fc70c0c026096556ef44fdfcbb48a43b9bc6ceef324dfe1122e868db57","0xC168E40227E4ebD8C1caE80F7a55a4F0e6D66C97",false,false],
 ["polygon","dhedge_dao_token","DHT",18,"dHedge DAO Token","304402206da3cefa573162edfe52a66dd7418c6f5014647cc970879a94ff83200def0bbf022043a610d9a386092babadd87b3aafb82c919d3631c19a89310cc10d16c3e11051","0x8C92e38eCA8210f4fcBf17F0951b198Dd7668292",false,false],
 ["polygon","dogelon","ELON",18,"Dogelon","304402204b0eb245902417b53d92071ea313cf7fbf4423bcc554c8a317676e35c12b964b0220078a990d4cbf28bd53a0a7a64252f8f074f678e6d6af10e8740faa577ebeeac7","0xE0339c80fFDE91F3e20494Df88d4206D86024cdF",false,false],
 ["polygon","dovu","DOV",18,"DOVU","3045022100f367c43817b181ca597d3eb32297cab50f90973ce3dda7857d6f663a74a514f40220055e8d0606c5d36b5c25c62c7d70cf8bb009bf8a8ab101612424d0ca3859574d","0x0E0ffc562d72316b783E887bbAAe1FD794ADb530",false,false],
 ["polygon","dragon_quick","DQUICK",18,"Dragon QUICK","3045022100ce96a6d8016442db7cb288d06e9df774d8615f40c4a8cd90c639c665d1f1bc5b02207acb92af880448671857d8a43508ac8aed68eb809c5ce2ae9a449bb6dd0bc930","0xf28164A485B0B2C90639E47b0f377b4a438a16B1",false,false],
 ["polygon","dsla","DSLA",18,"DSLA","3044022029ae2e5d5f851834c388141564c60a9d26c53bc81e9c6bf887f1206a86371b8f022075b60ffdf098243d8feab6ec0f55eca8c1722c1d7bf9a4c27cfcd44c439b557a","0xa0E390e9ceA0D0e8cd40048ced9fA9EA10D71639",false,false],
+["polygon","ecowatt_token","EWT",4,"Ecowatt Token","304502210090dd86bd0463a48c3b947b1e3bc08ebf3631202c676a857c5fc10525c8b7a65e02207b80b742250dcab35088fa81fc150ff998e3c070cb0bb455c11425e56ddcca52","0xd9E838dd60c8ea1e7dD4E670913323bB87DB112c",false,false],
 ["polygon","endor_protocol_token","EDR",18,"Endor Protocol Token","3044022007b3f74047928a24b8bffc08998b542d2631e354788be568e669bbe402eb474902201660b60dedbded95b83607adbad5f20a6e6f83a5c3070ef7c932b8b49b417249","0x1Ac417f1dD23504aAC2076D854581f03196403Fd",false,false],
 ["polygon","energy_web_token_bridged","EWTB",18,"Energy Web Token Bridged","30440220765fbb44728292f2fbc4c486ad1e12cd657d9772328bda64a1f64d8258a3bea302202042c8bc7cca6c63f55b4e596ccd3ac0279901a0fd8f7a0d0173fea7f2e8d5f3","0x43E4B063F96c33F0433863A927F5bad34bb4B03d",false,false],
 ["polygon","enjincoin","ENJ",18,"EnjinCoin","304402205eb28eb20df347c98e9d337feadd95b1cf96a6eeec81d65e607b2edc84e0a96c022011a4215ea48f0675496059434cc2f78bc1cfaf99049476081434a180e05700c6","0x7eC26842F195c852Fa843bB9f6D8B583a274a157",false,false],
@@ -70,6 +94,7 @@ module.exports = [
 ["polygon","fegtoken","FEG",9,"FEGtoken","30440220723ec9773b3cc9e6c3ce465c2b06acdb0b3382c445ff3ad263000e370219936c02204453a1c7ba6b266535f8651af12ad969a7744170f1ecd942a4263afec69b51cc","0xF391F574C63d9b8764B7a1F56D6383762E07B75B",false,false],
 ["polygon","fetch","FET",18,"Fetch","3045022100f462e081f273bb1ae69168d64d538ec202cffb609c7bfd306d12297e84825f1902207e82b440a4deeddf67292873c416ea065ef9663961a3ebfda099538f75d3c1d9","0x7583FEDDbceFA813dc18259940F76a02710A8905",false,false],
 ["polygon","fish","FISH",18,"Fish","3045022100b5c3f2cc9c6292ef20bac4d17e9e2cb4047a478cf75354817dd1fb6ac3fc7edd02202eba60971eb6987b3ee658c637cf905ec995600b3104beb7d97e03cb71010c90","0x3a3Df212b7AA91Aa0402B9035b098891d276572B",false,false],
+["polygon","flare_token","1FLR",18,"Flare Token","304402200a1e2ddbfbc9d4e948fa507ed54bc72d38d64c1ab9885527734a8c8836dc016b022062afbe992e01cf401a3c914e958983c55fb357db3228bde2c17771793ed3f3cf","0x5f0197Ba06860DaC7e31258BdF749F92b6a636d4",false,false],
 ["polygon","foam_token","FOAM",18,"FOAM Token","3045022100b66fe2c72a4073cad992f5c1f345fb1d4f057b79a4097fb913238e1e3f433785022006f6059384a3fd45623b8bdbd2514f3cebed6343d2548f3f34a2b5a2be39ebf9","0x66e16D50c07A01BB473eC794349d45aa1a0E5Dc2",false,false],
 ["polygon","fox","FOX",18,"FOX","30440220344a0fcdbff02568d465b09e2cffc047eb9169f9f89243cad21d20f78ab5749602206cee295f9945b6f10f984541b6975df9397edf35a3c9a0feba5d93fed87b04e0","0x65A05DB8322701724c197AF82C9CaE41195B0aA8",false,false],
 ["polygon","fractal_protocol_token","FCL",18,"Fractal Protocol Token","3045022100eb5ba2c5be243b68d60ae5c7295b79c491fe8b358d91186a81be9049136bc463022034399e2223bd6718aa718c5e29abee9f7c8035fb369ab50eb05f1856d852cc68","0x6f934b45fc6043527d73C7c99313dd838E723DB7",false,false],
@@ -77,10 +102,13 @@ module.exports = [
 ["polygon","frontier_token","FRONT",18,"Frontier Token","3044022031e7a7961b55db287bcc987825a632348a952dc9352abae8db1b0f3fe3b7353c0220404f9ea73f5dd8a9498fd6db2a69d5a4eb4dcc4847d0f3344567b6548676edbe","0xa3eD22EEE92a3872709823a6970069e12A4540Eb",false,false],
 ["polygon","furucombo","COMBO",18,"Furucombo","30440220126d341ad85e5d34cd11de5a0bd47e4e8ea494cd73af13b8b898aafd5bcd426b0220102a177dc04dd8db0200a5bf5f9607d898dc0dc39f74c23e163dfb685e897195","0x6DdB31002abC64e1479Fc439692F7eA061e78165",false,false],
 ["polygon","fusion","FSN",18,"Fusion","30440220777ef5482278d859c93f673444999f875168b22330e8a17323361f918217e64e0220351d14a816fb7a1d50839653927ed8d50d9d3084b496040a79085b6a2cb06808","0xFA1171334CB3a0f0A91E8cA6765F10e9638D1cBF",false,false],
+["polygon","gains_network","GNS",18,"Gains Network","3044022100eece560275b0df892c526a4248e43a00bc82eedfa958f5211f737e2572077ff7021f35ce82649c501c384f6af2d3372456d93991497021371ab34afbdc3230e45c","0xE5417Af564e4bFDA1c483642db72007871397896",false,false],
 ["polygon","gemini_dollar","GUSD",2,"Gemini dollar","3045022100a5715074876b0c47c69ebaa15cdc95cd449ddbf41378f2325b78587bb48b233002200c7192a6d3fb6f5e6aeeea99788a38ad951110fa9fa5d0bf44225612660168b5","0xC8A94a3d3D2dabC3C1CaffFFDcA6A7543c3e3e65",false,false],
 ["polygon","gitcoin_(pos)","GTC",18,"Gitcoin (PoS)","304402202c3d09cc0981d58ca9ec2ae33a571534f97b76e4cc6e012ab640e51d474df5b9022006c0b957028109d2c0b0130e11300b6653bbf169b08564bd162790d1b0adfff5","0xdb95f9188479575F3F718a245EcA1B3BF74567EC",false,false],
 ["polygon","gnosis","GNO",18,"Gnosis","30440220676c4fa426085479f0502e7c3d9bbeac273e356254ae3ba78bdf61efd225d8540220649a944cc30eb5ce55cb2d2899be899981716db2807514619d00adfd9fb8f3a0","0x5FFD62D3C3eE2E81C00A7b9079FB248e7dF024A8",false,false],
+["polygon","goblank_token_(pos)","BLANK",18,"GoBlank Token (PoS)","304402203bf51e0aeffc5c1887e1ea06043f5d015a6ef7cc0e0824aed1161f2f28b23f6f02200c5a55fd9887aebf11e77b889f245b53a715d07123ff1a60014e9c4fc536fd1b","0xf4C83080E80AE530d6f8180572cBbf1Ac9D5d435",true,true],
 ["polygon","golem_network_token","GLM",18,"Golem Network Token","3045022100effedc8a8cdfbe49bbdd207000529e9e56e35b5f29b677515c1dca0f715ec68e02203cb1c92a1704b6f6b64f75a161599e48183dc16e83de2b25ed12c1e32a53a992","0x0B220b82F3eA3B7F6d9A1D8ab58930C064A2b5Bf",false,false],
+["polygon","governance_abi","GABI",18,"Governance ABI","3045022100b6ba564715ca0a21f0c5e8c82f4d2aed7c76950ebee20b2e274983983a33692b022071d9edaa85cdce5e4dab4f9c572fa61c5f8a028f273206c554ff8d15000989bf","0xEd6AAb1615AaC7BC4C108dFd4Fdc9AD0c8304d47",false,false],
 ["polygon","graph_token","GRT",18,"Graph Token","304402206e352ac6947f2d2089339e567429a4de41f8c0ef4d0a7d4c0fb0a706a56e45af022011a85b67e644478abee9ed8e3f8e9db3973d54a85a9ab43d10f970be76a0d59b","0x5fe2B58c013d7601147DcdD68C143A77499f5531",false,false],
 ["polygon","gravity_finance","GFI",18,"Gravity Finance","3045022100bae69c4008de6e60a5a10a93639d003596660d587fc53da23050e71beebaac94022046ce5c5a776410d2ca615dc7914c86c903f2c39a23a5f0baf0da4972f28cf9aa","0x874e178A2f3f3F9d34db862453Cd756E7eAb0381",false,false],
 ["polygon","grumpy_cat","GRUMPY",9,"Grumpy Cat","30450221008a6a2a7ad22cef65134d40fdc80ea5166e249f96cf2a2d09aa47ef0cb5989cd202207786ae41332e7970ca184ffb012a3fafe37f16cf6b1123b481f8da17188a09fa","0x733726968Ae55dC58c26dBDc193FEa256a704ba4",false,false],
@@ -92,7 +120,9 @@ module.exports = [
 ["polygon","hopr_token","HOPR",18,"HOPR Token","3045022100ec6ee1c4afcf263626e01dabdd68d0c27a2830e8bc844f19dc7d5c8d01c5e846022031b10b45f3ea9c93074dcde99adf86f2042d298bddf0107f9ba473d11adea49c","0x6cCBF3627b2C83AFEF05bf2F035E7f7B210Fe30D",false,false],
 ["polygon","huobitoken","HT",18,"HuobiToken","304402202cc3a20b918bc1fb944b573207f4e8025b44b10bbef93c66c27cfcf22f526261022057132c7905fec1d6450034867a53f4c57ea4db958ff70755eb080f840caf6aaf","0xFAD65Eb62a97fF5Ed91B23aFD039956aaCa6e93b",false,false],
 ["polygon","husd","HUSD",8,"HUSD","3045022100df67ae2679473df7d0a7c7f4a77fdec16418988169d636f68c6c4355ddb0d66902203b6d8724bcb64fe4fb8513d46ae518abdd3ec88a6037d082a435e3a4d899e5b8","0x2088C47Fc0c78356c622F79dBa4CbE1cCfA84A91",false,false],
+["polygon","ig_gold_(pos)","IGG",6,"IG Gold (PoS)","3045022100dbbe3018e2a4f3ccf05bd069f252f6a567fc32ebb60940a194d67c1de817f437022014a67ea4bf5a3945467d2b2924d46023bd742ae68ba07decc1c2f786d0dc0dc0","0xe6FC6C7CB6d2c31b359A49A33eF08aB87F4dE7CE",false,false],
 ["polygon","impermax","IMX",18,"Impermax","3044022057045fbdcb5de6009d1a0b0a982363e38d2f1bdb48a383da30f52da897f544d002207fdd46c2b42b4b1447fb1b831e9f75bafd1ad21c09e0c584e5e55f58ea125609","0x60bB3D364B765C497C8cE50AE0Ae3f0882c5bD05",false,false],
+["polygon","impermax_borrowable","IMXB",18,"Impermax Borrowable","304402204b52cacbce56ddfc68c1a2549535512bf9bd27f4b9c816a782a6515fce5d92f1022039484a3ff1da5e9203b39dd1e8f08f0b8afaf9b0ddae526cacc160815b2c6caa","0x6d24195060e20C2d717408B8dEcCA487B945c78b",false,false],
 ["polygon","injective_token","INJ",18,"Injective Token","3045022100901789cfe2edb9719c1b3297f10c15dc3b33e7778118f2c731027795bf386c7702202db85f2c221e575fd3eeebc3a55273de7d820c2f14011bfb1ec2bdea2d372ec2","0x4E8dc2149EaC3f3dEf36b1c281EA466338249371",false,false],
 ["polygon","internxt","INXT",8,"Internxt","3044022017c8702e4eb6de4412fe11ef9bb79b465a3a652ca9a19e8a8b9e176a5ad432a002206413dc14ad5b40e55c9dfd407e16ae768b040f15a272ce4cfb916b5d08bda2ef","0xE64106154816A252da7728E35A0060F15d66cb34",false,false],
 ["polygon","inverse","XIV",18,"INVERSE","304402202c2ed3446f1a9d10e6691bf6ab73fd2172335642b4d358fda5253229ee70d18c0220399d76cf157f660673f2abaac5daaad0953d14b2ad681ef84a79c18e883ee270","0xde799636aF0d8D65a17AAa83b66cBBE9B185EB01",false,false],
@@ -100,8 +130,14 @@ module.exports = [
 ["polygon","iron_finance_ice_token","ICE",18,"Iron Finance ICE Token","304402205ba575a2e5ec8428b5833ef20673f1a5b72ca974fb8d672ba24d592ca160b24602206eab86bbdb0aae8bfcbfc9d9c298e63a6e1e7831f82ffbf19575a1f8aacc6609","0x4A81f8796e0c6Ad4877A51C86693B0dE8093F2ef",false,false],
 ["polygon","iron_stablecoin","IRON",18,"IRON Stablecoin","3045022100cdf97a354a7a3d78bd1e9f4fad24b7f795004ed8f8542d310a04ce9c511899bc022044f68b145028da1ce6f3ef32f3c89e0c1e943b24e403e410b43d134f3e2b7c06","0xD86b5923F3AD7b585eD81B448170ae026c65ae9a",false,false],
 ["polygon","iron_titanium_token","TITAN",18,"IRON Titanium Token","304502210086601ee708d443615e629b377537bb649b3c327b49faa654daccf232e2ab22cd0220357ac0f5cd3866211d0f90cf590f838cbb881929e8138fa67ed6ba68e9576b41","0xaAa5B9e6c589642f98a1cDA99B9D024B8407285A",false,false],
+["polygon","jamonswapv2","JAMON",18,"JamonSwapV2","30450221009011b6d63f6a6d09a50aa581fe3f25589f84b7cef0163d89f302adcc714ff86f02202752c4a6a1889d734ef28f167e5bed8f5417f3722282b0f73f8fd4baedea0351","0x2594C5B25EdAb3eBe937e27650c02922D9A62BF2",false,false],
 ["polygon","jarvis_reward_token","JRT",18,"Jarvis Reward Token","3044022064f387e73f3549e57f3d3e491d891050bb22f3fbba60b50bccb05dd269a9dfae02200e7c59234ce6ae72b1c5495f14ae289b07eca39e28e168bf83cc2672e4559e31","0x596eBE76e2DB4470966ea395B0d063aC6197A8C5",false,false],
+["polygon","jarvis_synthetic_canadian_dollar","JCAD",18,"Jarvis Synthetic Canadian Dollar","3044022013bcad79311c2f66aab3fc4f760269f8d9b3d73b505acaaa6d6330e5ff83aadb02207a746067202ccb8bfece4a97973cb53cdca0becec02200ed7bd0bc5a2060cda7","0x8ca194A3b22077359b5732DE53373D4afC11DeE3",false,false],
+["polygon","jarvis_synthetic_euro","JEUR",18,"Jarvis Synthetic Euro","3044022011607ee079738049f3a5c3213ce9282960420bcf139d249ef5d34fe46a54970f02203967a87e580e152c01ab5a4145a61a33cedac08a6d97fcb3c6df3516ec5388e0","0x4e3Decbb3645551B8A19f0eA1678079FCB33fB4c",false,false],
+["polygon","kart_racing_league","KRL",18,"Kart Racing League","3044022007d27359e5a392b47f7ebd60a9741d19c9d0272f607b6cb7f6cff00a816810950220205025d3e774c634540c65cbcb4acea2ca8b24bf0bc41c900e51bf9c74490a32","0xc17b109E146934D36c33E55FADE9cBDa791b0366",false,false],
+["polygon","kastatoken","KASTA",18,"KastaToken","3045022100de98dbd09128c748cf82459c1bebdd57e86387bdf88a50f14745195bff26747c022002036319474361dbeb6af529a51e8885d95c0e95612aaedf5731a8836b2e1b48","0x235737dBb56e8517391473f7c964DB31fA6ef280",false,false],
 ["polygon","keep_token","KEEP",18,"KEEP Token","30440220421592d52f76c4a708667d808924c882743f4e158529f11ba66acccf012cb84a02207c34642ab8380c5c3ddaa0ccb112b35483c41c06678589e09c1e79dc54636a16","0x42f37A1296b2981F7C3cAcEd84c5096b2Eb0C72C",false,false],
+["polygon","kommunitas","KOM",8,"Kommunitas","3045022100b24917d3aa3d7d34599ed683335c1b75e7cb7ec88994e41cb41f91be2fad9e960220334bba59f53368d880ba2cef7119cfa3b77f75371abe6dfe3060e466d3139039","0xC004e2318722EA2b15499D6375905d75Ee5390B8",false,false],
 ["polygon","krill","KRILL",18,"Krill","30450221009145d9d117de32047ffd5acc07f236b67dd59529b9fe9c3adb9e3ac0233ba0300220514c59fcc843eb3d4820d374dd17d3a415c33b4f40ff192e4c6b686ad615f932","0x05089C9EBFFa4F0AcA269e32056b1b36B37ED71b",false,false],
 ["polygon","kyber_network_crystal_v2_(pos)","KNC",18,"Kyber Network Crystal v2 (PoS)","3045022100ac428da56dd621eaaaafc550ec9970ac2c8c5297af8bc7bcff068f35890121010220779cdf971ea6b145fcca46cdf8a926acbaa9ff97dd309ed5770cf769e7cdaa3d","0x1C954E8fe737F99f68Fa1CCda3e51ebDB291948C",false,false],
 ["polygon","launchpool_token","LPOOL",18,"Launchpool token","30440220215b5326f5901a6105e12af79906312949c5f8b004386edb10f4cbc903ba720a02205182105c2768f6ea61e18d3731cc50daf40e1eab935f3d94cfd135fb95146eb8","0x9e7Deba360c77949bb75d16DC81b2393C15005Eb",false,false],
@@ -109,18 +145,27 @@ module.exports = [
 ["polygon","litentry","LIT",18,"Litentry","3045022100a4d37029bb8562a35730ce18e8f2f9acff6c52d3fb2f67dd41bfc72b97390074022053a10ad4cfba695e88596bdca5d3ce6dfed8a596546c6f3e90a531e9ca237301","0xe6E320b7bB22018D6CA1F4D8cea1365eF5d25ced",false,false],
 ["polygon","livepeer_token","LPT",18,"Livepeer Token","3045022100998fd2d3d64f163248ac84d8751d447edd362dda82ce83efd6ca7cbee8ef4e4f02205a78aff434d95ee987acb6823dd2281603141af99b287482ca8f2b04e337449d","0x3962F4A0A0051DccE0be73A7e09cEf5756736712",false,false],
 ["polygon","loopringcoin_v2","LRC",18,"LoopringCoin V2","304402200d9794985a79f6bcdfbd8284811d80249ca32467f9ef66a27c0fcdff43225c88022008e2692ad59eaf726882c8505fa2cec99451060181707abd41abb30565826881","0x84e1670F61347CDaeD56dcc736FB990fBB47ddC1",false,false],
+["polygon","lucky_races_sliver","SLIVER",18,"Lucky Races Sliver","304402205764d60f07714fb0c49ac04013453164a90074c60239403a09a421d70f5bf56902206c5a83cdcad134a95b1959c7fc98850fddee0b0eccade2568e1d65154f489d83","0xA95b410743e8c2F7b64f1373D3Ca2B3454864a94",true,true],
+["polygon","mac&cheese_token","XVMC",18,"Mac&Cheese Token","3045022100eddbabb393f94fede1c3f0dd764f4921b089ae8fcfb3bc753d1be46da59507f0022058b822bbe5bdb52b18bcc883a75752c3f9d94148483043bfc5110776ee3afa4a","0x6d0c966c8A09e354Df9C48b446A474CE3343D912",false,false],
+["polygon","magicstone","MST",18,"MagicStone","304402200530a9ae82b7124e48f766bc53a4a5b7e1618b137206f7d320e9e1f2a318c8650220054be0aee9698591b3b89224ab63f7422b9db1f17ce69707916b940dc45bfd63","0xa353dEb6Fb81dF3844D8bd614D33d040fDBb8188",true,true],
 ["polygon","mainframe_token","MFT",18,"Mainframe Token","3044022050a25d9d1399fdda062949b6c8eec25b2144e39db8aa0e8103e69da78c68b7da02202459fe33547637f812422f7f397e19fe3b363107ad64b2cd8ef9c0b540ecda90","0x91cA694d2B293f70Fe722fbA7d8A5259188959c3",false,false],
 ["polygon","maker","MKR",18,"Maker","3044022028163f6351dc91fb1b53d02e407b7d41e7d43eebe30e9b51ba26c227ab1447f302202257ef7960134df342c1a74c06698e5f0cb2180032dad663ad17ff6b850f4588","0x6f7C932e7684666C9fd1d44527765433e01fF61d",false,false],
 ["polygon","mantra_dao","OM",18,"MANTRA DAO","30440220610533443b38a8a6567b13823a5e34275edf5532e0ccb28678b4571aa485428e02202f41793d82e11b5736dd936611854e6352eb08ddd5017e9b3f036e7cc66d8333","0xC3Ec80343D2bae2F8E680FDADDe7C17E71E114ea",false,false],
 ["polygon","marlin_pond","POND",18,"Marlin POND","30450221009f09ff2b458405be622481bc389a31906218473270a91c855103294b705e855d02204eebc0115274f969e0890fb3f6ea239a04af5e9a5ab4cb4545cb514c0910ee2a","0x73580A2416A57f1C4b6391DBA688A9e4f7DBECE0",false,false],
+["polygon","marshall_rogan_inu","MRI",18,"Marshall Rogan Inu","304402200e865057ca5e07054e690db8d866052f9921fad73d6c5333b0d093dc5e0c44810220269b8b1b5e95b628c68bd3c36f47193f34cd0c9d8dd96caf8271d266cbf7aff2","0x0913dDAE242839f8995c0375493f9a1A3Bddc977",false,false],
 ["polygon","mask_network","MASK",18,"Mask Network","304402201610449bb6144e5fa5a0dff8a3937546c63f5425c3183040bc3af91d241fb10d022050f2e11ab90ba3355081db37e2e1391eb670458095a5f960a498746239bc887e","0x2B9E7ccDF0F4e5B24757c1E1a80e311E34Cb10c7",false,false],
 ["polygon","matic_aave_interest_bearing_usdc","MAUSDC",6,"Matic Aave interest bearing USDC","30450221009485660477820f0c20fe0152f8697419d74b86ae006571da3a82b92df08a771502201d43918816b858ca2afd797634b4061c5fea30d7689cec8aa37df8701ce782d5","0x9719d867A500Ef117cC201206B8ab51e794d3F82",false,false],
 ["polygon","meta","MTA",18,"Meta","304402200deefaa443552d4634507754bf388eab461a5be5489fcaac40795ada4bfc366702203b2fbb7076432cbdaff88002e616f80de97841426b9a3cf824fe2bf57035c7db","0xF501dd45a1198C2E1b5aEF5314A68B9006D842E0",false,false],
+["polygon","metaverse_index_(pos)","MVI",18,"Metaverse Index (PoS)","30440220542c66ae5d89e71f53d7bf85b47651947dca243f56153804cee0d3a06eab7398022040b0bca5debb473b88a0019724e10a259d18c66246eaa86d83d02747419ac03b","0xfe712251173A2cd5F5bE2B46Bb528328EA3565E1",false,false],
 ["polygon","mimatic","MIMATIC",18,"miMATIC","3045022100e90762884ff53d623b79c5bafc7b6ef9b15dbce9a0b2a6cb9606ef41d8c9f967022030aad9cd5db83ed7ab378fcfae58b462c52c147b0cc432304fb822e4edf0cef2","0xa3Fa99A148fA48D14Ed51d610c367C61876997F1",false,false],
 ["polygon","moonwolf.io","WOLF",9,"moonwolf.io","304402207219cde47ae2f296ba82fa38d0170c7a28002339e39e362138c4acd5d8c53c46022016ceebddc7101938b8ac86f21b16cc2366f95c5a50b2b79e98f896a75e30b2fe","0x8f18dC399594b451EdA8c5da02d0563c0b2d0f16",false,false],
 ["polygon","mover","MOVE",18,"Mover","3045022100c481954a8735c5670354f4355213d85f4145739fdb5084ee33ec448bbedfc634022012023de7dcfe7f37fcad39506ec865738bae14a12c15d31b0241250c42ceec3e","0x521CddC0CBa84F14c69C1E99249F781AA73Ee0BC",false,false],
+["polygon","mysterium_(pos)","MYST",18,"Mysterium (PoS)","3045022100b3c724140250803ab3294d03eef28097363b46446913a382990608994b7f6df40220255f5f5b4c23cbe16f8d32a28f5cae94c2f4e81eb0372e87bf884b02221c1271","0x1379E8886A944d2D9d440b3d88DF536Aea08d9F3",false,false],
 ["polygon","napoleonx","NPX",2,"NapoleonX","3045022100948dd459eb87f596ee1555f6be889bba96ff1764a2648e43c24720f0e3fd7d8f02205b3100d2b087d3555954788b1991cb4543535f84954bdccfa3a87a5fc700c35a","0x106035DEDcA635d75d6Ff05F98F25916829eDB77",false,false],
+["polygon","neighbourhoods_token","NHT",18,"Neighbourhoods Token","3045022100fe6ee0005f410db9901de157784de9e58469bb2d957c8c219bea4f41eb57a495022044fbb21830ba0ef9ff2aaa0c991702f802cb6b95248e6c297e21b15d9ca77705","0x84342e932797FC62814189f01F0Fb05F52519708",false,false],
+["polygon","neos_credits_(pos)","NCR",18,"Neos Credits (PoS)","3045022100bdd5489d4dd7cf302678e96fc09c757a7abf7f8480f08bfe39fe2032982a8f9202205d9f97b690d49f46444dd434bb95064702d91b65f5a234516c062dbd7e46263c","0xDA6f726E2088F129D3Ecb2257206AdF7D8537Ba5",false,false],
 ["polygon","nexo","NEXO",18,"Nexo","3045022100e55f2f5e587f95f627932523c95ea0d1d748dbecc9921d0def3f0acccb70efee02207c3b2fee173e99d90c28eb324832db0d14b4a477c1ec5f2468442b6404d41d8c","0x41b3966B4FF7b427969ddf5da3627d6AEAE9a48E",false,false],
+["polygon","nextearthtoken","NXTT",18,"NextEarthToken","3045022100e58ea45d8190eda21e1283f7690a367f8b70cd1864e086cbd24236fd448149a9022049a3ee879b9bd229602ff2b0f59736be95abd53ef6921c2aba60b25facadf894","0x0d0B8488222F7f83B23E365320a4021b12eAD608",false,false],
 ["polygon","occ","OCC",18,"OCC","3045022100ae2ad9e38419d203873ef47a3450a0ff1cefcc02db4820ba2d78eb366af1b828022018e0d61c5904170cc32090826efbaccfa813988a2043588fd70badaa8bb1b975","0xF30355BA14B2d3b31597EC71a4542f57e902Cb80",false,false],
 ["polygon","ocean_token","OCEAN",18,"Ocean Token","3044022042dbec9c2f3a5270e89008659bcd0d3103922f210f9e4d574e189690690ece5c02207a959c67010943e050dd27a9a01f1362bb82a4aae92ed011edc748215edc115e","0x282d8efCe846A88B159800bd4130ad77443Fa1A1",false,false],
 ["polygon","oiler","OIL",18,"Oiler","3045022100d677ed4bffea3fb7bbbd7ff8e0c5b7e5a0a4da19492b5467cc08998446454e9002206da6c0c70ad320c4e03f132d6a65a24887d757f2e6cb6c92387fb5760bdfb818","0x65DfA90171e1bd9060209405026C4E1f4A8E58df",false,false],
@@ -136,10 +181,12 @@ module.exports = [
 ["polygon","parsiq_token","PRQ",18,"Parsiq Token","30450221008443ea75db0a8bbfecaa35a7867697ef097e9a09ccafd7e23742804e5bd5a92102207feeac4b67416298ec7a4b6288a111e8e60cbdbe0413832ddb5dd992ccfd77f4","0x9377Eeb7419486FD4D485671d50baa4BF77c2222",false,false],
 ["polygon","paxos_gold","PAXG",18,"Paxos Gold","30440220366777a5244a25b29267cf4d4df4d285d3f288205aa25e453c9746315c04681c02206df03e2f914dd3561716e4fe1f8a9a5b2e4a59a4291dd777feaddcb3d660959e","0x553d3D295e0f695B9228246232eDF400ed3560B5",false,false],
 ["polygon","paxos_standard","PAX",18,"Paxos Standard","3045022100a0fd20a02a801d127eb9e6ad59856909ab580c6ba2ee7d8757705e4335f16e2802204949c6795e2e23b11d07f78a045b42900195f1b3d3599357aa9058a84e8058ba","0x6F3B3286fd86d8b47EC737CEB3D0D354cc657B3e",false,false],
+["polygon","pegaxy_stone","PGX",18,"Pegaxy Stone","30440220431b2054344c423673cdd3ded15c3a5ae82dcd1e2683b8ced95fcef042a4660f02205d1ba3b5e9fa516a49abea1477639b62c09a3380fadf3d4964abb87327761c73","0xc1c93D475dc82Fe72DBC7074d55f5a734F8cEEAE",false,false],
 ["polygon","pepedex","PPDEX",18,"Pepedex","304402200b590012ea3cb13b39348b11863fc7073f54ec13b94b5345d2d021695ea7056402207c7990ae61a9e4cbac791896d57da4609180f5738c1d806c27ffeb09ea251e7e","0x127984b5E6d5c59f81DACc9F1C8b3Bdc8494572e",false,false],
 ["polygon","piedao_dough_v2","DOUGH",18,"PieDAO DOUGH v2","3045022100f08450a344629b75ff4f4ac08a21cea60603a46c0c2795e2782b770180838de602205a50961be1e93f6b71b91a3444e70c13c22ed37315c0767e45bca8835c874554","0x990F860968714c2D16e91Ec0cd9709A94264Cf64",false,false],
 ["polygon","pinakion","PNK",18,"Pinakion","304402204c37665d5ae80c17bb4710b7db4e924a8479675c8255637566b8ebb56d052dc902204091c0317a8497f4267eb1c70d3e953cc9332785f59a912ddf80cb6e91af8fd3","0xad93E067e149f0A5ecd12D8EA83B05581dD6374C",false,false],
 ["polygon","pinknode_token","PNODE",18,"Pinknode Token","30440220351b833f6ba2a58c10d9bf92a04dd0e507bb0ec1bf0cc228da153e98a648324e022070025b9a70ebf6a9e8367dcda3ee5d9f1325f585516d51627a60e9875cb48bf4","0x4E92f6f10496F3fCB790Cc6969Cc15B87c8F9199",false,false],
+["polygon","playermon","PYM",18,"Playermon","3045022100e9ae502812883cb8bce68153548f5302e6ea41483bbfed576f8435802198b2f402205c1c7727151f5f5b6c4e7290bef11aa3dca95f8496aa04873dd5ec17b2c0ef55","0x0bD49815EA8e2682220BCB41524c0dd10Ba71d41",true,true],
 ["polygon","pluton","PLU",18,"Pluton","304402207f4935970038b11ffc3bee1567913d40f1ae3f30471d752a8531ba795a89cd51022035938e84b0edf51bb21d820eb561d08567dbe8918b396136e5afdc9102b7ef14","0x7dc0cb65EC6019330a6841e9c274f2EE57A6CA6C",false,false],
 ["polygon","polydoge","POLYDOGE",18,"PolyDoge","3045022100932c4b2dd1a5e77023209930d92f84cf27d9629b75f8368434760cd2bde68924022013d2b965abc44f75bb5c7d91609a123d5dd050d4ed0ff6fac4ae9cf8a00aeb29","0x8A953CfE442c5E8855cc6c61b1293FA648BAE472",false,false],
 ["polygon","polygaj_token","GAJ",18,"PolyGaj Token","30450221009f4e2e496ba2ec107711f8299f17cff60df06aac4f746ac5d8b2af63411db53202202b1bc6868e97589d9881575e323553e3b6f0392f374bf57d6444e6ebffb605f8","0xF4B0903774532AEe5ee567C02aaB681a81539e92",false,false],
@@ -154,6 +201,7 @@ module.exports = [
 ["polygon","pup_token","PUP",18,"Pup Token","304402206a285914d15826aae6434475ae938c43063a4d92aa1784609ce54de24f9cea6b02207881f884d2220550021c997a0df85e83c38c1e04add8ba2de9e994d5129e3b6a","0xcFe2cF35D2bDDE84967e67d00aD74237e234CE59",false,false],
 ["polygon","pusd","PUSD",18,"PUSD","3045022100bfa2a989bbe0efcf2c8da49b356da55c00b4b434afcfe4eedef863ccff92df4f02205d6ff069bcc8c60d2c891528498667057223ed65ca26867b8aaceb991d113d19","0x9aF3b7DC29D3C4B1A5731408B6A9656fA7aC3b72",false,false],
 ["polygon","pyq","PYQ",18,"PYQ","3045022100fd7084feafd0b7aa463ec98d8e3df78c2aa0c661d56b6d0e1b247dd9fb0cffdb022040b00978f591d18e2d2f197c1d511709302ae970a2d20eadbec36fdfdb254eb8","0x5a3064CbDCCF428ae907796cF6aD5a664CD7F3d8",false,false],
+["polygon","pyr_token","PYR",18,"PYR Token","304402201f9a2c5b19732e43d8e09701aa2969e5ee1b0394665cecb558dd1d0afa962bc4022057cceedec8dd7fad2299463f50f91b1e098b4d6f3f864bd1ae5fec5ea63ef36a","0x430EF9263E76DAE63c84292C3409D61c598E9682",false,false],
 ["polygon","qi_dao","QI",18,"Qi Dao","304402206efdc615e0a058701d70bd924a858d8cf38214d1136a170e19a680ae31395f23022049c98bf1df2c5d7d2d1739060c0554f8631e278919b3a1e739246bd848c2bdb7","0x580A84C73811E1839F75d86d75d88cCa0c241fF4",false,false],
 ["polygon","quickswap","QUICK",18,"Quickswap","3045022100c098673b45df93df9cf10141bf5e43c673dc0f8b5d5ccd9f1e9bebea6a8ba022022000e645464709f47288a7b38e30054094854a898efb478fb31cc2971564d7ef6e","0x831753DD7087CaC61aB5644b308642cc1c33Dc13",false,false],
 ["polygon","rarible","RARI",18,"Rarible","304402201d63d1eca3dcc61f0c155bd0cdb34c5cd7c2ac446a64d8a25c2fce5dfc0ff11a022012243245f3ce383d22c2bf8a164f23f9b59c1ea9ce8a9abaf681df4dcc3e30d6","0x780053837cE2CeEaD2A90D9151aA21FC89eD49c2",false,false],
@@ -161,33 +209,42 @@ module.exports = [
 ["polygon","render_token","RNDR",18,"Render Token","3044022042d62a1be583b33bf641857e3519ef7422a350e21258f030b8c25fdd4c7e8260022052f4e10c82849f947f4a5acb87db379c56c84576bd083073c16a6a59869096cf","0x61299774020dA444Af134c82fa83E3810b309991",false,false],
 ["polygon","rendoge","RENDOGE",8,"renDOGE","30450221009459770c7f112f22fe117f69f315d68f97fe0e4627415732e4ff57e9f768052602206a26ff549a74a8c5c40e114bb19e3f2cf30c0f7dec91d5745148ff218b5236ff","0xcE829A89d4A55a63418bcC43F00145adef0eDB8E",false,false],
 ["polygon","renfil","RENFIL",18,"renFIL","3045022100bb66e93c3aa7aead7ef944c20614e837aa963d064330d0b98ead27c915101c1702204fee135f515103754423dd144d822efadceb848489508c2477150969a88242e1","0xc4Ace9278e7E01755B670C0838c3106367639962",false,false],
+["polygon","revv","REVV",18,"REVV","304402203d883d329adb570698e893c5ad8f83b3716e89f5e8749bac4a742c83f741e5ca02203884a0cf71fdedbb30d41ec293722cc39e89b2b31fd9d0ac4363020c39bbcaf1","0x70c006878a5A50Ed185ac4C87d837633923De296",true,true],
 ["polygon","ripiocreditnetwork","RCN",18,"RipioCreditNetwork","3045022100b5b8c036b1906e9f286ac8ef6ccf20bdbd944c22476b1a4fc7a8cff541cbfd85022060c5a569ca87042188efaeb352fac5cd41a9f0416aa304e713684232cb009ed1","0x89C296Be2F904F3e99a6125815b4B78F5388d2dd",false,false],
 ["polygon","rlc","RLC",9,"RLC","30440220763bbbb5486d36f9d455764c8632008b1c5b67d45c099e3ad492b8ecbbd290f5022007ff2d836f1668768f9abde172de2a7e5d11299f12a91b1172e17b0f4263c610","0xbe662058e00849C3Eef2AC9664f37fEfdF2cdbFE",false,false],
 ["polygon","rocket_pool","RPL",18,"Rocket Pool","304502210088e2753ca251b934fb95645f451b30f0d69b72228f76e671c71f4784d80523df0220695ecb3258a48d7fc434e73910e1718b20a6359e4501c796d541337a2c820a79","0x7205705771547cF79201111B4bd8aaF29467b9eC",false,false],
 ["polygon","route","ROUTE",18,"Route","304502210089f5b01d29803b7a94d038092c05f2378ed2fb73a4bccb8a08fec6e75f43836302206851dfb9b543f033484550b1041bdc9bed00d43a03b70bc3d312c8edd94c8314","0x16ECCfDbb4eE1A85A33f3A9B21175Cd7Ae753dB4",false,false],
+["polygon","rubic_(pos)","RBC",18,"Rubic (PoS)","3045022100f849984f462557a6b74b48337387eaff6439cba4e90242e958d91d5e3c88e5a30220740585ce2bec064bbc8c9a4a77b1bd7c923ffabf97b0f1ee6535237d759985c4","0xc3cFFDAf8F3fdF07da6D5e3A89B8723D5E385ff8",false,false],
 ["polygon","safedollar.fi_dollar_2.0","SDO",18,"SafeDollar.Fi Dollar 2.0","3045022100b9cd7e36d080fa92721d5ea291caa453784a3b84070fe0387ad5018792a0648a022028b80062c9598c4977b9c2e5e908371599801732ee8341f55c0758d51f692d3d","0x66C59Dded4EF01a3412a8B019B6e41D4a8C49A35",false,false],
 ["polygon","safedollar.fi_share_2.0","SDS",18,"SafeDollar.Fi Share 2.0","304502210084d9c869d4bc12ca8a6653a736c52284fbf9503bba4659acc52161f7db97213f022056fcad1644a1cac28d21a905c0376550cb430105dea74112b841e5a20b2564bf","0xAB72EE159Ff70b64beEcBbB0FbBE58b372391C54",false,false],
-["polygon","sand","SAND",18,"SAND","30450221008c73c747568e70852fc34ee10b589a92aa0185b226ce2608eeea5436e55b1d36022030ae32c2b9ab0100597156083e3fb9c7f07f51422a01fa1289f9080b0c5c07b6","0xC6d54D2f624bc83815b49d9c2203b1330B841cA0",false,false],
+["polygon","sand","SAND",18,"SAND","3045022100c21119cb079c76295c8524b3091064854b02ff371416a1a5e48e9095edb3bfe502206f785716e63e80a99eade8d3fb273031614393e086883d866b5f9db2809b9ebc","0xBbba073C31bF03b8ACf7c28EF0738DeCF3695683",false,false],
+["polygon","sand_old","SAND_",18,"SAND_","3045022100e49d3eb141bdd203f2c1b85d351540ab0632be7a398a39498182387c4adc3875022027f9a9c5f5927727682d71e7224828dbc79926991cee0521501c840f8cb2e57d","0xC6d54D2f624bc83815b49d9c2203b1330B841cA0",false,false],
 ["polygon","sapien_network","SPN",6,"Sapien Network","3044022077e4b88efdb063bfaea7f1cd49b1e3698c7eec3f700889a34f2df8c5e6dc5f44022036ccfac493a01ef8268aff6b295633b36ec5df127c74daf35d8c3ea2eae91f01","0xeAb9Cfb094db203e6035c2e7268A86DEbeD5BD14",false,false],
 ["polygon","sentinel","SENT",8,"SENTinel","304402201f023fd294b88242f14e4061fe352aee3aeceaf28d47c3989f1ea7aa97f2aa0302206856a4013b36c60d7a0fdcba49e43f007d252cddc25cecad7bcaf71e59eab064","0x48e3883233461C2eF4cB3FcF419D6db07fb86CeA",false,false],
 ["polygon","serum","SRM",6,"Serum","3045022100c649349ccb7fa669ae7fcb0397f145d7ac72f1fb472842b28d838b250595c9ef02200b67a2fb4e51dee5b542d74d62dbea0892e0fc59e9a0c02772670cf83f327056","0x6Bf2eb299E51Fc5DF30Dec81D9445dDe70e3F185",false,false],
 ["polygon","shopping.io","SPI",18,"Shopping.io","3045022100a62733824ef978d69bf87b0b58cf62781c4dec1986e88c79a9ce90252e1a48c902202ce77ea5f7759fe688619f0a8bec2528b459c715e47c3d43636e3d94f2c9c39c","0x068180071617528606371c31892EcBF2B70ac1d2",false,false],
 ["polygon","singularitynet_token","AGIX",8,"SingularityNET Token","3045022100c8522f1c1ab31af32f71740b443c3cc1039c74764c59d4ea15cad27c4097410802202a2dacb662495b79a7573dad329ae177d40bc37976d94a9d4ec9c1eedd21c7e7","0x190Eb8a183D22a4bdf278c6791b152228857c033",false,false],
 ["polygon","spacerat","SRAT",9,"SpaceRat","304402201ed3b017ce531c52a7f9df50ec38610d8463b7729e44511f1bf0f58d07bab49402201af4319aecd3c33a0209e355e195b2cc20fff1292b39f0aef67a10d0d2be35cf","0x1132f58810Ee9fF13E97aECCd8DDa688Cc5eb8F4",false,false],
+["polygon","sphere_finance","SPHERE",18,"Sphere Finance","304402201115976dd774b1cef894ecfe78d69738f91eb5d64d4a8ba57a76320c2a83f9c302202cce636c5bd99c6e7d214633666a3584faf4961a36c72f0f9f45bf4ccbf6548f","0x8D546026012bF75073d8A586f24A5d5ff75b9716",true,true],
 ["polygon","stacker_ventures_token_(pos)","STACK",18,"Stacker Ventures Token (PoS)","3044022011c1caac8801e8a3209b04f8516f2618339ac18f9ccf2438f9793545772caaed02207956082348e52ab602bd61bc732c93d18a7b288d9b5c20a4bb9309876be72a6a","0xccBe9B810d6574701d324fD6DbE0A1b68f9d5bf7",false,false],
+["polygon","stackos","STOS",18,"StackOS","304402206addcc7967322c1123edc8c40c15612734bc224b98c3cfd73c29a53aa8971fb4022031fc00cb726d7f6274213e26c04cb3a4f931df3814dc8350f6dd37bb559a0768","0x980111ae1B84E50222C8843e3A7a038F36Fecd2b",false,false],
 ["polygon","stake","STAKE",18,"STAKE","304402205f7a4020031482ade4568896f97a0575a3dc1919646e7d8b4b414739d0847c4a02200d254c84b384ec61376d83d53a7b2d14bd246acbce7a8d3234636108fb3c1322","0xeb5c9e515629B725D3588A55E2a43964DcfdE8cA",false,false],
+["polygon","staked_abi","SABI",9,"Staked ABI","3045022100a9918d3d13f424bffcaeb646051b4f7ff3dc48b103ead0bf39d3de0e84ce29c302202bac77bb63eb94b2b946fe9ddaf63c36dcd277be86974ebdeed551b9430f9ae0","0x925a785a347f4a03529b06C50fa1b9a10808CAb5",false,false],
+["polygon","stakeshare","SSX",18,"StakeShare","3045022100bc6a6b01d812191c0c9fd8d5936ada0a2230ce6d0b7cdc1d31271f5340f9f4fc02201694ec8737dc0f7d320fc811a6d6d5f077f84490f8aab578e1551af360fca37d","0x9E2d266D6c90F6C0D80a88159b15958f7135B8Af",true,true],
 ["polygon","stakewise","SWISE",18,"StakeWise","3045022100cbb1f417533c99ff0635b3b5d79de8af2f58c81dce21c45889f5d558cce33bf4022049ea64d7339b8f65c1fdc22dce407a2bf3e39f90c08713019eacb3a42d58e2ee","0xBbbe128568222623D21299f019a978C8587B33dC",false,false],
 ["polygon","stasis_eurs_token","EURS",2,"STASIS EURS Token","3045022100cd91f0b64f1981b6c541efa3d28c1eefd0df08cde9c385b6eebc5a1d76725f2e022047f17ddc307ed6941857f5cab5acbd5f1bebc3551a987d8614342c7379825672","0xE111178A87A3BFf0c8d18DECBa5798827539Ae99",false,false],
+["polygon","streamr","STREA",18,"Streamr","304402201d7ab4d608b973717f1fd75e3f728ada6d9afeea8f76765c61ee233a7368e3b40220734224fecc9aeb414c4b396175ec54bcb124b6437448f84f6a1c9f5499a8a202","0x3a9A81d576d83FF21f26f325066054540720fC34",false,false],
 ["polygon","student_coin","STC",18,"Student Coin","30450221008682480e8ac8a19f23dc4d21ef7a6701d663e0b047e9315bb214695515b076fb0220295e4854bb0862e5aed648cb7fd4967d0998e2631ceb23897e5c2983048513e3","0x779b299eA455D35A44FE9baC48648be22C08DEa2",false,false],
+["polygon","superfarm_(pos)","SUPER",18,"SuperFarm (PoS)","3045022100b34b5b5981c20518320baf7c6f695fda645162a870dccfc292a32ef12c163b5d022065681bcaf2a7018a3638dfd880526b84418b22e323d590d564198e501f730000","0xa1428174F516F527fafdD146b883bB4428682737",false,false],
 ["polygon","sushitoken_(pos)","SUSHI",18,"SushiToken (PoS)","30450221008884a0addca753dc4393d8e5541819547c6d6e5ead02a5e1a7235fcddbb0c64902204c5c0f309402efa8767c9114ccf2b866d2c0a2aa05cb7d14aa18e00bf841a508","0x0b3F868E0BE5597D5DB7fEB59E1CADBb0fdDa50a",false,false],
 ["polygon","swipe","SXP",18,"Swipe","304502210091117e42be90ced27fecb5c5061e7924163c8fa2aa5f97ff015591825ff034bb022036fb7411de65f0cf2c3f657b40b8426447b5d3ac5f2d7f17e1f02d7a705834b5","0x6aBB753C1893194DE4a83c6e8B4EadFc105Fd5f5",false,false],
 ["polygon","swissborg","CHSB",8,"SwissBorg","3045022100b336701b18926b318036944930d9bc8961785d5b8bff98108909e99ca63a54410220687e186ba97314ef15a15d3b753d4162240505a2a6730b65a30d96b5645a4c79","0x67Ce67ec4fCd4aCa0Fcb738dD080b2a21ff69D75",false,false],
 ["polygon","synth_susd","SUSD",18,"Synth sUSD","30450221009977e9457dbc4e54cbeac056370046e91aba5cce6b2f4ed218c520404a352f4502203b23604adeab58c71558dcb261054b3bdbe16aa29f58c767807765ee1b3b3673","0xF81b4Bec6Ca8f9fe7bE01CA734F55B2b6e03A7a0",false,false],
 ["polygon","synthetix_network_token_(pos)","SNX",18,"Synthetix Network Token (PoS)","304402201f2c8892ac4f7edda2054cc124893abebf1b1341609c310b010c24f3eb5d53ab02205622033d1f4ded85a0de1061b3e65fc707718029f02b75d56bfe470e6061661b","0x50B728D8D964fd00C2d0AAD81718b71311feF68a",false,false],
 ["polygon","tapmydata","TAP",18,"Tapmydata","3045022100f0507c1386c9af0892dfc5a7fd72b880126ebaa601b0c528655dbe2f434cabbf022018de2cb1bbb5d67d633e25429573ed9210ae5b8466acdf64a858d23a9511fd38","0x10635bF5c17F5E4c0Ed9012aEf7C12f96a57a4Dd",false,false],
+["polygon","tattoomoney","TAT2",18,"TattooMoney","30440220618d64e47a74b6b398341005a1ea578c72789f45f74351655c9b2fd435a53452022078ab85fec2bd2f10513770e1bee7f36c6c9503f4b8865e5d0f9d774596a15586","0x2596A8B90D39EB89d5668ca5B6deE54E9ae3d4C0",false,false],
 ["polygon","telcoin","TEL",2,"Telcoin","3045022100df6734e8dc79763f3a2e3ccbbe547f9380f24a5d25ec0e7ef5e3357aa23e105b0220484abf1214cdc1bda02a17900fc9f3ae73ffbddbebb101f51ad5674c860d4aa2","0xdF7837DE1F2Fa4631D716CF2502f8b230F1dcc32",false,false],
 ["polygon","tellor_tributes","TRB",18,"Tellor Tributes","304402207d9872134e96a0deebacd719b1a3d684a8a3282449d721ae0bcdfdf1a864f6e402200c07a0aa5e0d897fdc795ccc9794e5e288b8e22e30594b25f8afbacb053b4370","0xE3322702BEdaaEd36CdDAb233360B939775ae5f1",false,false],
-["polygon","theta_token_(pos)","THETA",18,"Theta Token (PoS)","3043021f4639b4633adfc6abe3be6f10dbabd381bc2ca6855eacf7ca6e407889575f4f02203d9608c0beefb31108259088234f6c55d3933b0813d784cb583af7351a748c64","0xB46E0ae620EFd98516f49bb00263317096C114b2",false,false],
 ["polygon","tidal_token","TIDAL",18,"Tidal Token","30440220347494aaf6a8f7b0bb0d7fc44955c6d9be79f3b969a5032a1a94eaadedf4fefe022052ffbf3c709f8ea5840f68e014edb071b234563d901965ee69a4db5a4f5b4b42","0xac311fA85E8645bB1DB53ECc04Fa8d41997FACF8",false,false],
 ["polygon","torn_token","TORN",18,"TORN Token","30440220707fcc4100917670748fd1d0268837866914590842aae8ccda38a1c4f06f996c02206d66134328fecd3d9a19c560e102f5eb870030c87a2c65ae55a9b9a79665ee1e","0x23fE1Ee2f536427B7e8aC02FB037A7f867037Fe8",false,false],
 ["polygon","trueusd","TUSD",18,"TrueUSD","3045022100eef61cb3c069df526c50ffd331b25fdccff35a1e6485047b00617ad316e305cc0220235b4b5259c38256e11d6cf9710b4ec6ab02aad6509754a5e8cd63c7bab9a56f","0x2e1AD108fF1D8C782fcBbB89AAd783aC49586756",false,false],
@@ -196,17 +253,26 @@ module.exports = [
 ["polygon","unibright","UBT",8,"UniBright","3045022100ea250710f1e7f25a56fd4abac52ee893c8c09099ec56e7918c1146bb381b511d02205243299925580b29209f9de2d85dcb7a3d0b35358330cf55cf75434b8d5d8d3a","0x7FBc10850caE055B27039aF31bD258430e714c62",false,false],
 ["polygon","unifund","IFUND",18,"UNIFUND","30450221009b57ca04ac3d25d14d1bae394ba30c0fa459154f7dfe702953db535d78c4287a0220440a23085a1004f725bac0d9e20d279ef854b14c27379bb71df1c02e47d7de64","0x4ba47B10EA8f544F8969BA61dF3E5bE67692a122",false,false],
 ["polygon","unimex.network","UMX",18,"unimex.network","30450221009a031e72a2716c560623d166433722ea1ddc68739be51dfabc80d9b18dc604fc022076849337d463345ddabe847a931aa220ca6677e1d4170fd7491a90e960980282","0x5dF1a47e05b4caBa0ED3Df13662642f05314764D",false,false],
+["polygon","uniristoken_(pos)","UCO",18,"UnirisToken (PoS)","3045022100ac7931fe9057e3f8189962b11e4c05e195bc1565c4f9e702c5942f92179b5590022022b063abf2426f010f6d9d395260a389aae4da0b2856a241c1e071ffd7f246d8","0x3C720206bFaCB2d16fA3ac0ed87D2048Dbc401Fc",false,false],
 ["polygon","uniswap_(pos)","UNI",18,"Uniswap (PoS)","304402206fd1ec530e75eadcd8adca249ec060cec1ee532d2a63c6617de8ec73cc62456502207d2e780bb93cca35ea810596b764cbe8fa72a593379ff4f48512687cc4611fd3","0xb33EaAd8d922B1083446DC23f610c2567fB5180f",false,false],
+["polygon","uniswap_v2","UNIWEWM",18,"Uniswap V2","3045022100ea8aadc21ac89ff39521db07d8e1d438f3085e7455ae2aff9565b5faf2cf573e022072493a0d257545f3cda3fe6d4922a01fef131a1ffd2eb1c5a760c5bb1b05b9ef","0xadbF1854e5883eB8aa7BAf50705338739e558E5b",true,true],
+["polygon","uniswap_v2_","UNIWEUSDC",18,"Uniswap V2","30450221008340415d778cff8a08d547eed2ff628523e59759d5ed14e19768f15f5bdad1fc02203283dad240ae0da57df3a62869af97b31833db7741f235a2ee2838e86192a4b1","0x853Ee4b2A13f8a742d64C8F088bE7bA2131f670d",true,true],
+["polygon","uniswap_v2__","UNIWMUSDC",18,"Uniswap V2","3045022100ead796edd11d1451ca745a495b92b00e193b5798d038e274c31c04de90059f0e02201bc8be5b7b1151ee55e1f0061d44285e075a2c408c1063e09c5f0935fae22829","0x6e7a5FAFcec6BB1e78bAE2A1F0B612012BF14827",true,true],
+["polygon","uniswap_v2___","UNIWMUSDT",18,"Uniswap V2","304402205c8d7db793ff7e0208eb885646b28bf4a0a8d94036dfcfc60749d03fe183492302200fe71aafae326ade9464ccaa6691a221177e06ecce1dfaa817888718b79b502b","0x604229c960e5CACF2aaEAc8Be68Ac07BA9dF81c3",true,true],
 ["polygon","usd_coin_(pos)","USDC",6,"USD Coin (PoS)","304402200ef170e3f0a71b03ad7758e201d9861102f78cabcbbc5dbfecc5a946236438c7022043c591264f57cf3026f12f98e9b5112d2dd2e7b2d6c04206ace2475ed21e80f4","0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174",false,false],
 ["polygon","usdk","USDK",18,"USDK","304402205613ac257b797ef31c6890956a12d192b89d03a7a2078898dafb0b79b0cb1b5c02202dfe6eff014166937d540742538335257dd9f5f9dce2491b7c05dbbbe92f4080","0xD07A7FAc2857901E4bEC0D89bBDAe764723AAB86",false,false],
 ["polygon","vespertoken","VSP",18,"VesperToken","3044022046419ec25dab4fc07fe8a0936cea029dfc04ce9cc9955a7d1b16c5f6361a3b940220336259779ffbec502fe8a318101def15008d5077e8ed88a0af1ea8db15bee2a0","0x09C5a4BCA808bD1ba2b8E6B3aAF7442046B4ca5B",false,false],
+["polygon","vigorus","VIS",18,"Vigorus","304402200ca0d53d4f7d8c01fcd949130b112eac9070a2eb19641a2f1a6d81c49574d6420220647a7b7e9680002069fa2827951458493f3b3cdc22bdd50f228c85e937694700","0xcC1B9517460D8aE86fe576f614d091fCa65a28Fc",false,false],
+["polygon","voxel_token","VOXEL",18,"VOXEL Token","3045022100e9df989b25b7dab962f5fdfbce74cb79fd51a566fb656b82cdae730b2dac2a3602204854f994fb4dbd20dfe8346f047d971dc5ef19d8648e53b8f27cd908e6943d46","0xd0258a3fD00f38aa8090dfee343f10A9D4d30D3F",false,false],
 ["polygon","voyager","VGX",8,"Voyager","304402207adb14a4402bd7f0e74ed94c4e0212785b145e671f2ae9b888ef6fa1e85cddd7022070f8f3ea98dd86ae7434cece3acf132dc6f0a34b4e51c439c03807e4abefc580","0x054c42b6414747F5263b4A86f21B1aFAD00326Bf",false,false],
 ["polygon","waultswap_polygon","WEXPOLY",18,"WaultSwap Polygon","3045022100bc695093f7bc1f4a5eca39fb0ad45bb7d3421fb24c0b582456e7a3e6556cfeff02207aa4ef9f0b705bbc5ee20fc26a27d8292a21ddf5c3b1f679bbe224d54246397c","0x4c4BF319237D98a30A929A96112EfFa8DA3510EB",false,false],
 ["polygon","wepower","WPR",18,"WePower","3045022100b2022385b9fb931b51eb9dc58b6acfd9329dc01fd876143fca05dd13f1a00072022033c14a9f00ce8b5ca8618f64bde930921e09af636208776f3a0520a247747508","0xCBAb27B2D5b468Eea509Bd430100922341BC3f1A",false,false],
 ["polygon","woofy","WOOFY",12,"Woofy","3045022100a708d0d98a979ed4407844ee000dc827f2812baf88abb3b547f644c610a467ae022064c0c24a928ef14c212a73ebffb566150861a026ebb503a2886e331766fb1fa8","0xfce22AD802Bbc9FE74534575b166D06198bD66ea",false,false],
 ["polygon","wootrade_network","WOO",18,"Wootrade Network","304402206fa91ec7507c0b90e8072900b0ec42a60496deefa710cb0a942178610021100f022040c899804634ebce099a9c8ffb3bd1e34a229e73a4190c0f320662fda1ee0994","0x1B815d120B3eF02039Ee11dC2d33DE7aA4a8C603",false,false],
 ["polygon","wrapped_celo","WCELO",18,"Wrapped Celo","304402204533f1d264f27aaf2bded35c70e342f9e093d3563e21a69581099694b76296630220224c28490a748c183cda281bada3a27958f561fdc33f71e998d62001a21f4e1f","0xc30DD0cCE2e683aA68c87C280e73beB3DEc9B3BA",false,false],
+["polygon","wrapped_ether","WETH",18,"Wrapped Ether","304402207b674dc939fc9958aef413f7e1b3d755ffbcc8240933d7b6d888e4adeae5c8a802206dd9b1c0c8291e28977497ac8139902446fbb770cd82f6aad7eb7c7b72d69b50","0x7ceB23fD6bC0adD59E62ac25578270cFf1b9f619",false,false],
 ["polygon","wrapped_filecoin_(pos)","WFIL",18,"Wrapped Filecoin (PoS)","3044022078b0af0f86bc7b62ca450d4868649cf8df6973d3331a7ea244abec015abbbf5c02204a80853c3be77db55d88d647b8a1522defa0a0ea6a30482ca815453d180451d4","0xEde1B77C0Ccc45BFa949636757cd2cA7eF30137F",false,false],
+["polygon","wrapped_ixs_token","WIXS",18,"Wrapped Ixs Token","3044022010c5a4c283d33754c01d52b1efb06e45844dd575d1bb20ad164af0eb915d0f3a022060075f9a9a4c8c19339931a68114ff2dc85ee8057cbac439eb2e7803463a4c67","0x1BA17C639BdaeCd8DC4AAc37df062d17ee43a1b8",false,false],
 ["polygon","wrapped_matic","WMATIC",18,"Wrapped Matic","304502210083ae7e7508bbe06a1b6c59543e62a3dcb2e825a6c39c1adc9ccd439253ac7f4d02201f60dcd93ad430b79b4b0b39571f5f15988233c04e5d2e45a4ecce1639fabf07","0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270",false,false],
 ["polygon","wrapped_ust_token","UST",18,"Wrapped UST Token","304402204ad3702836823f826e934b945997b15c3238887767be33ef21f1aa5c9337c95d02206ed2ae004976c4d4ef8d6d62a785602f620490ad36b447d08de0aed9e46a4a02","0x692597b009d13C4049a947CAB2239b7d6517875F",false,false],
 ["polygon","xtoken","XTK",18,"xToken","3045022100eb4c86fd40f2cb862f560b6f2c39a7e20cd7756ccee500e5a71105d7ebecfec002203f445753745434711b27ad0470b5f407933742a90a1abfc0a63d332606e91361","0x22308317A4aA9e8196065E0b3647e582d44c4d7b",false,false],
