@@ -1,17 +1,22 @@
 import axios from "axios";
 import { getLoadConfig } from "./loadConfig";
-import type { LoadConfig } from "./loadConfig";
-import { log } from "@ledgerhq/logs";
+import type { LoadConfig } from "../types";
 
 type NftInfo = {
   contractAddress: string;
   collectionName: string;
-  data: Buffer;
+  data: string;
 };
 
 type BackendResponse = {
   payload: string;
 };
+
+function axiosErrorHandling(e) {
+  if (!e || !e.status) throw e;
+  if (e.status === 404) return null;
+  throw e;
+}
 
 export const getNFTInfo = async (
   contractAddress: string,
@@ -24,19 +29,25 @@ export const getNFTInfo = async (
   const response = await axios
     .get<BackendResponse>(url)
     .then((r) => r.data)
-    .catch((e) => {
-      log("error", "could not fetch from " + url + ": " + String(e));
-      return null;
-    });
+    .catch(axiosErrorHandling);
   if (!response) return;
 
+  // APDU response specification: https://ledgerhq.atlassian.net/wiki/spaces/WALLETCO/pages/3269984297/NFT-1+NFT+Backend+design#NFT-Metadata-BLOB
   const payload = response["payload"];
-  const collectionNameLength = Number(payload.slice(2, 3));
-  const collectionName = payload.slice(3, 3 + collectionNameLength).toString();
+  // Collection name length position: 3rd byte -> caracter 4 to 6
+  const collectionNameLength = parseInt(payload.slice(4, 6), 16);
+  const collectionNameHex = payload.substr(6, collectionNameLength * 2);
+  const collectionName = collectionNameHex
+    .match(/.{2}/g) // split every 2 characters
+    ?.reduce(
+      (acc, curr) => (acc += String.fromCharCode(parseInt(curr, 16))),
+      ""
+    ); // convert hex to string
+
   return {
     contractAddress: contractAddress,
-    collectionName: collectionName,
-    data: Buffer.from(payload, "hex"),
+    collectionName: collectionName || "",
+    data: payload,
   };
 };
 
@@ -48,15 +59,12 @@ export const loadNftPlugin = async (
 ): Promise<string | undefined> => {
   const { nftExplorerBaseURL } = getLoadConfig(userLoadConfig);
   if (!nftExplorerBaseURL) return;
-  const url = `${nftExplorerBaseURL}/${chainId}/contracts/${contractAddress}/plugin/selector?selector=${selector}`;
+  const url = `${nftExplorerBaseURL}/${chainId}/contracts/${contractAddress}/plugin-selector/${selector}`;
 
   const response = await axios
     .get<BackendResponse>(url)
     .then((r) => r.data)
-    .catch((e) => {
-      log("error", "could not fetch from " + url + ": " + String(e));
-      return null;
-    });
+    .catch(axiosErrorHandling);
   if (!response) return;
 
   const payload = response["payload"];
