@@ -12,6 +12,23 @@ import { identifyUSBProductId } from "@ledgerhq/devices";
 import { CantOpenDevice } from "@ledgerhq/errors";
 import { listenDevices } from "./listenDevices";
 let transportInstance;
+
+const DISCONNECT_TIMEOUT = 5000;
+let disconnectTimeout;
+const clearDisconnectTimeout = () => {
+  if (disconnectTimeout) {
+    clearTimeout(disconnectTimeout);
+  }
+};
+
+const setDisconnectTimeout = () => {
+  clearDisconnectTimeout();
+  disconnectTimeout = setTimeout(
+    () => TransportNodeHidSingleton.autoDisconnect(),
+    DISCONNECT_TIMEOUT
+  );
+};
+
 /**
  * node-hid Transport implementation
  * @example
@@ -21,6 +38,7 @@ let transportInstance;
  */
 
 export default class TransportNodeHidSingleton extends TransportNodeHidNoEvents {
+  preventAutoDisconnect = false;
   /**
    *
    */
@@ -89,6 +107,20 @@ export default class TransportNodeHidSingleton extends TransportNodeHidNoEvents 
   };
 
   /**
+   * convenience wrapper for auto-disconnect logic
+   */
+  static async autoDisconnect(): void {
+    if (transportInstance && !transportInstance.preventAutoDisconnect) {
+      log("hid-verbose", "triggering auto disconnect");
+      TransportNodeHidSingleton.disconnect();
+    } else if (transportInstance) {
+      // If we have disabled the auto-disconnect, try again in DISCONNECT_TIMEOUT
+      clearDisconnectTimeout();
+      setDisconnectTimeout();
+    }
+  }
+
+  /**
    * globally disconnect the transport singleton
    */
   static async disconnect() {
@@ -97,12 +129,14 @@ export default class TransportNodeHidSingleton extends TransportNodeHidNoEvents 
       transportInstance.emit("disconnect");
       transportInstance = null;
     }
+    clearDisconnectTimeout();
   }
 
   /**
    * if path="" is not provided, the library will take the first device
    */
   static open(): Promise<TransportNodeHidSingleton> {
+    clearDisconnectTimeout();
     return Promise.resolve().then(() => {
       if (transportInstance) {
         log("hid-verbose", "reusing opened transport instance");
@@ -138,8 +172,26 @@ export default class TransportNodeHidSingleton extends TransportNodeHidNoEvents 
     });
   }
 
-  close() {
-    // intentionally, a close will not effectively close the hid connection
+  setAllowAutoDisconnect(allow: boolean): void {
+    this.preventAutoDisconnect = !allow;
+  }
+
+  /**
+   * Exchange with the device using APDU protocol.
+   * @param apdu
+   * @returns a promise of apdu response
+   */
+  async exchange(apdu: Buffer): Promise<Buffer> {
+    clearDisconnectTimeout();
+    const result = await super.exchange(apdu);
+    setDisconnectTimeout();
+    return result;
+  }
+
+  close(): Promise<void> {
+    // intentionally, a close will not effectively close the hid connection but
+    // will allow an auto-disconnection after some inactivity
+    this.preventAutoDisconnect = false;
     return Promise.resolve();
   }
 }
